cache: separate cache code from shard code
Separate the Cache-related code to a new `cache.go` file, leaving just
the shard code in `clockpro.go`. This makes it easier to work with the
code.
