+++ internal/cache/value.go
@@ -41,7 +41,12 @@ type Value struct {
 // want to add finalizer assertions.
 const valueEntryGoAllocated = !buildtags.Cgo || invariants.UseFinalizers
 
-func newValue(n int) *Value {
+// Alloc allocates a byte slice of the specified size, possibly reusing
+// previously allocated but unused memory. The memory backing the value is
+// manually managed. The caller MUST either add the value to the cache (via
+// Cache.Set), or release the value (via Cache.Free). Failure to do so will
+// result in a memory leak.
+func Alloc(n int) *Value {
 	if n == 0 {
 		return nil
 	}
@@ -125,4 +130,14 @@ func (v *Value) Release() {
 	if v != nil && v.ref.release() {
 		v.free()
 	}
+}
+
+// Free frees the specified value. The buffer associated with the value will
+// possibly be reused, making it invalid to use the buffer after calling
+// Free. Do not call Free on a value that has been added to the cache.
+func Free(v *Value) {
+	if n := v.refs(); n > 1 {
+		panic(fmt.Sprintf("pebble: Value has been added to the cache: refs=%d", n))
+	}
+	v.Release()
 }
